Improve test code
