+++ micrometer-tracing/build.gradle
@@ -7,6 +7,5 @@ dependencies {
     itImplementation project(":testing")
     itImplementation "io.micrometer:micrometer-tracing-bridge-brave:$micrometerTracingVersion"
     itImplementation "io.micrometer:micrometer-tracing-bridge-otel:$micrometerTracingVersion"
-    itImplementation "io.zipkin.brave:brave-instrumentation-kafka-clients:5.15.1"
     itImplementation "io.opentelemetry.instrumentation:opentelemetry-kafka-clients-2.6:1.26.0-alpha"
 }
